re-org: rename package name from env to config
