+++ core/src/main/java/com/linkall/vance/core/http/HttpServerImpl.java
@@ -16,7 +16,7 @@
  */
 package com.linkall.vance.core.http;
 
-import com.linkall.vance.common.env.ConfigUtil;
+import com.linkall.vance.common.config.ConfigUtil;
 import com.linkall.vance.core.Adapter;
 import com.linkall.vance.core.Adapter2;
 import io.cloudevents.CloudEvent;
